It looks like you have updated the `install.sh` scripts in your Git repository to use the correct user `$_REMOTE_USER` instead of `vscode`. This is a good change as it ensures that the commands are run with the correct user when the container is launched from VSCode's Remote-SSH extension.
The changes you made affect the following scripts:

* jupyter-environment/install.sh
* lazygit/install.sh
* memgpt-environment/install.sh
* mkdocs-environment/install.sh
* opencommit/opencommit.sh
* setup/install.sh

These scripts are used to install various features and dependencies for your project when the container is launched. By updating them to use the correct user, you ensure that these installations are done with the correct permissions and in the correct location on the remote server.

Overall, this looks like a positive change to your Git repository. It ensures that the container is launched correctly with the correct user and that the installations are done with the correct permissions.

 It looks like you have made changes to your Dev Container configuration files, specifically in the `features` folder. The changes include updating the Terraform environment manager (tfenv) feature to use the remote user's home directory instead of the default "vscode" user and updating the Conda feature to use the remote user instead of "vscode". Additionally, the Vimrc feature has been updated to use the remote user's home directory for installing plugins and themes.

These changes should not cause any issues as long as the remote user has sufficient permissions to execute the commands and install the required software packages. However, it is important to note that if you are using a shared or public environment, these changes may introduce security risks if the remote user's home directory is accessible by other users.

It is also worth noting that the `wget` command used in the Vimrc feature installation may not work reliably in all environments as it relies on having a working `curl` or similar library installed. In such cases, using alternative methods like cloning repositories using SSH or Git protocols can be more reliable and consistent.