It looks like you have added a new devcontainer feature called `azure-cli-persistence`. This feature aims to preserve the `~/.azure` folder across container instances of Azure CLI, avoiding the need for extra logins.
The feature includes a `devcontainer-feature.json` file that defines its metadata and configuration options. It also has an `install.sh` script for setting up the cache directory and symlinking it to the user's home folder. The `oncreate.sh` script is used during container creation, and it checks if the user is not root and chowns the `/dc/azure-cli` directory to the user if necessary.

To use this feature in your devcontainer configuration, you should include it as a dependency in your `devcontainer.json` file:

```json
{
  "name": "devops-toolkit",
  "image": "docker.io/robinmordasiewicz/devcontainer:latest",
  "initializeCommand": "docker pull robinmordasiewicz/devcontainer:latest",
  "runArgs": ["--name", "devops-toolkit", "--hostname", "devops-toolkit"],
  "features": [
    { "id": "azure-cli-persistence" }
  ]
}
```

Make sure to create the `.devcontainer` folder and place the `devcontainer.json` file inside it before using your devcontainer in VSCode.