Added FortiDevSec extension to devcontainer.json, devcontainer-feature.json and install.sh files.
The FortiDevSec Extension is a comprehensive SaaS-based tool for software developers and DevOps to find and fix all types of application security issues within their CI/CD cycle. This extension has been added to the list of features in the devcontainer.json file, along with its corresponding devcontainer-feature.json and install.sh files.

These changes will enable the FortiDevSec Extension to be installed automatically when using this devcontainer.