It looks like there were some changes made to the `devcontainer.json` file, which is used to define a Dev Container for Visual Studio Code. The changes include adding a new feature called "fortidevsec" and removing an unused command from the "onCreateCommand". Additionally, it seems that the Docker image being used has changed from "mcr.microsoft.com/devcontainers/base:ubuntu20.04" to "mcr.microsoft.com/devcontainers/base:ubuntu22.04".
There were also changes made to the `README.md` file, which includes instructions for building and pushing a Dev Container image using Docker. The `concepts/index.md` file appears to have been removed entirely. Lastly, there were some minor formatting changes made to the `logo.svg` file and the `mkdocs.yml` file.

It's unclear what the purpose of these changes are or if they will have any impact on the functionality of the repository. Without more context it's hard to say for certain.